フロアの分割・使わないコードの削除
