+++ CESA-2020-Prototype/Assets/Prefabs/LiftFloor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3ff1ba480eaec30479318195d1a07d8e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
