+++ CESA-2020-Prototype/Assets/Prefabs/MoveFloor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 09cd1dda4f23299448227fffe4a8dd07
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
