+++ CESA-2020-Prototype/Assets/Prefabs/RotationFloor.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4ad08a7333f7f6b4f8bd536dcd33cf97
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
