+++ CESA-2020-Prototype/Assets/Scripts/Stage/Floor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f97da022d4c6c264fb477abc7024c65d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
