+++ CESA-2020-Prototype/Assets/Scripts/Stage/Floor/FallFloor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 687a28ec22c6a81478c69b67321dd51d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
