+++ CESA-2020-Prototype/Assets/Scripts/Stage/Floor/LiftFloor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e37ed5d7c936cd418924b990d111e7c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
