+++ CESA-2020-Prototype/Assets/Scripts/Stage/Floor/MoveFloor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 180bce267cbc7be4ba93d6c58416a4ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
