+++ CESA-2020-Prototype/Assets/Scripts/Stage/Floor/RotationFloor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6a4a9cabd40fb94e80c0d8454a50eca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
